Fix the minimum send and receive buffer sizes used in setsockopt.
In Linux, the minimum buffer size used in setsockopt for SO_SNDBUF and
SO_RCVBUF option is the value defined in SOCK_MIN_RCVBUF and SOCK_MIN_SNDBUF.
(These values depend on the struct sk_buff size).

gVisor was using minimum value in tcp_rmem/tcp_wmem in setsockopt. This CL
will fix this and use the value defined in SOCK_MIN_RCVBUF and SOCK_MIN_SNDBUF
in setsockopt.

Auto tuning will continue to use the values in tcp_rmem/tcp_wmem which is the
same as linux.

PiperOrigin-RevId: 419874904