Make the webhook deterministic.
We add a simple image that will generate certificates as a init container,
and refer to this directly use the Kubernetes yaml. Note that this also fixes
the issue of versioning within the yaml file, as this will now generate a
version of this file that refers to the precise local versions.

PiperOrigin-RevId: 420826312